+++ Controller/Admin/TransactionList.php
@@ -12,7 +12,8 @@ namespace Wirecard\Oxid\Controller\Admin;
 use OxidEsales\Eshop\Application\Controller\Admin\AdminListController;
 use OxidEsales\Eshop\Core\Registry;
 use OxidEsales\Eshop\Core\TableViewNameGenerator;
-use Wirecard\Oxid\Core\Helper;
+
+use Wirecard\Oxid\Core\PaymentMethodHelper;
 use Wirecard\Oxid\Model\Transaction;
 
 /**
@@ -53,7 +54,7 @@ class TransactionList extends AdminListController
     public function render(): string
     {
         $this->_aViewData += [
-            'payments' => Helper::getModulePayments(),
+            'payments' => PaymentMethodHelper::getModulePayments(),
             'actions' => Transaction::getActions(),
             'states' => Transaction::getStates(),
         ];
