+++ Core/OxidEEEvents.php
@@ -11,6 +11,7 @@ namespace Wirecard\Oxid\Core;
 
 use OxidEsales\Eshop\Core\DatabaseProvider;
 use OxidEsales\Eshop\Core\Registry;
+
 use Wirecard\Oxid\Extend\Model\Order;
 use Wirecard\Oxid\Model\Transaction;
