+++ Extend/Model/Order.php
@@ -13,11 +13,10 @@ use OxidEsales\Eshop\Application\Model\Country;
 use OxidEsales\Eshop\Application\Model\OrderArticle;
 use OxidEsales\Eshop\Application\Model\User;
 use OxidEsales\Eshop\Application\Model\Basket;
-
 use OxidEsales\EshopCommunity\Core\Field;
 use OxidEsales\EshopCommunity\Core\Registry;
-use Wirecard\PaymentSdk\Entity\AccountHolder;
 
+use Wirecard\PaymentSdk\Entity\AccountHolder;
 use Wirecard\Oxid\Core\Helper;
 use Wirecard\Oxid\Model\Transaction;
 use Wirecard\Oxid\Model\TransactionList;
