+++ Model/CreditCardPaymentMethod.php
@@ -12,6 +12,7 @@ namespace Wirecard\Oxid\Model;
 use OxidEsales\Eshop\Application\Model\Payment;
 use OxidEsales\Eshop\Core\Config;
 use OxidEsales\Eshop\Core\Registry;
+
 use Wirecard\Oxid\Core\Helper;
 use Wirecard\Oxid\Model\Transaction as TransactionModel;
