+++ Model/PaymentMethod.php
@@ -10,13 +10,14 @@
 namespace Wirecard\Oxid\Model;
 
 use Wirecard\Oxid\Core\Helper;
+
 use Wirecard\PaymentSdk\Config\Config;
-use Exception;
 
 use OxidEsales\Eshop\Application\Model\Payment;
 use OxidEsales\Eshop\Core\Registry;
 
 use Psr\Log\LoggerInterface;
+use Exception;
 
 /**
  * Class PaymentMethod
